# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Content Atlas API",
    "description": "A data consolidation platform for SMBs to consolidate data from multiple sources into PostgreSQL",
    "version": "1.0.1"
  },
  "paths": {
...
    "/import-history": {
      "get": {
        "tags": [
          "import-history"
        ],
        "summary": "List Import History",
        "description": "List import history with optional filters.\n\nParameters:\n- table_name: Filter by destination table name\n- user_id: Filter by user ID\n- status: Filter by status ('success', 'failed', 'partial')\n- file_name: Filter by original file name\n- file_hash: Filter by SHA-256 hash\n- source_path: Filter by stored source path (e.g. B2 key)\n- file_size_bytes: Filter by file size recorded during import\n- limit: Maximum number of records to return (default: 100)\n- offset: Number of records to skip for pagination (default: 0)",
        "operationId": "list_import_history_import_history_get",
        "parameters": [
          {
            "name": "table_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Table Name"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "file_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File Name"
            }
          },
          {
            "name": "file_hash",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File Hash"
            }
          },
          {
            "name": "source_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Path"
            }
          },
          {
            "name": "file_size_bytes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File Size Bytes"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportHistoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import-history/{import_id}": {
      "get": {
        "tags": [
          "import-history"
        ],
        "summary": "Get Import Detail",
        "description": "Get detailed information about a specific import.\n\nParameters:\n- import_id: UUID of the import to retrieve",
        "operationId": "get_import_detail_import_history__import_id__get",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportHistoryDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import-history/statistics": {
      "get": {
        "tags": [
          "import-history"
        ],
        "summary": "Get Import Statistics Endpoint",
        "description": "Get aggregate statistics about imports.\n\nParameters:\n- table_name: Filter by destination table name\n- user_id: Filter by user ID\n- days: Number of days to look back (default: 30)",
        "operationId": "get_import_statistics_endpoint_import_history_statistics_get",
        "parameters": [
          {
            "name": "table_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Table Name"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportStatisticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import-history/{import_id}/duplicates": {
      "get": {
        "tags": [
          "import-history"
        ],
        "summary": "Get Import Duplicate Rows",
        "description": "Retrieve duplicate rows that were detected during an import.",
        "operationId": "get_import_duplicate_rows_import_history__import_id__duplicates_get",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDuplicateRowsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/import-history/{import_id}/duplicates/{duplicate_id}": {
      "get": {
        "tags": [
          "import-history"
        ],
        "summary": "Get Duplicate Row Detail Endpoint",
        "operationId": "get_duplicate_row_detail_endpoint_import_history__import_id__duplicates__duplicate_id__get",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          },
          {
            "name": "duplicate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Duplicate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import-history/{import_id}/duplicates/{duplicate_id}/merge": {
      "post": {
        "tags": [
          "import-history"
        ],
        "summary": "Merge Duplicate Row Endpoint",
        "operationId": "merge_duplicate_row_endpoint_import_history__import_id__duplicates__duplicate_id__merge_post",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          },
          {
            "name": "duplicate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Duplicate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateMergeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateMergeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import-history/{import_id}/validation-failures": {
      "get": {
        "tags": [
          "import-history"
        ],
        "summary": "List Validation Failures Endpoint",
        "description": "List validation failures for a specific import.",
        "operationId": "list_validation_failures_endpoint_import_history__import_id__validation_failures_get",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "include_resolved",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Resolved"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportValidationFailuresResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import-history/{import_id}/validation-failures/{failure_id}": {
      "get": {
        "tags": [
          "import-history"
        ],
        "summary": "Get Validation Failure Detail Endpoint",
        "description": "Get detailed information about a specific validation failure.",
        "operationId": "get_validation_failure_detail_endpoint_import_history__import_id__validation_failures__failure_id__get",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          },
          {
            "name": "failure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Failure Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationFailureDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import-history/{import_id}/validation-failures/{failure_id}/resolve": {
      "post": {
        "tags": [
          "import-history"
        ],
        "summary": "Resolve Validation Failure Endpoint",
        "description": "Resolve a validation failure.",
        "operationId": "resolve_validation_failure_endpoint_import_history__import_id__validation_failures__failure_id__resolve_post",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          },
          {
            "name": "failure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Failure Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveValidationFailureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveValidationFailureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportHistoryDetailResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "import_record": {
            "$ref": "#/components/schemas/ImportHistoryRecord"
          }
        },
        "type": "object",
        "required": [
          "success",
          "import_record"
        ],
        "title": "ImportHistoryDetailResponse",
        "description": "Response for single import detail"
      },
      "ImportHistoryListResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "imports": {
            "items": {
              "$ref": "#/components/schemas/ImportHistoryRecord"
            },
            "type": "array",
            "title": "Imports"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "success",
          "imports",
          "total_count",
          "limit",
          "offset"
        ],
        "title": "ImportHistoryListResponse",
        "description": "Response for import history list"
      },
...
      "ResolveValidationFailureResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "failure": {
            "$ref": "#/components/schemas/ValidationFailureRow"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "failure",
          "message"
        ],
        "title": "ResolveValidationFailureResponse",
        "description": "Response from resolving a validation failure."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}